Rework on imports block of generated classes
