+++ generator/src/main/java/io/doov/gen/ModelMapGenMojo.java
@@ -17,8 +17,7 @@ package io.doov.gen;
 
 import static io.doov.gen.FieldInfoGen.constants;
 import static io.doov.gen.FieldInfoGen.createFieldInfos;
-import static io.doov.gen.FieldInfoGen.imports;
-import static io.doov.gen.FieldInfoGen.methods;
+import static io.doov.gen.DslMethodsGen.methods;
 import static io.doov.gen.ModelWrapperGen.*;
 import static io.doov.gen.utils.ClassUtils.transformPathToMethod;
 import static java.nio.file.Files.createDirectories;
@@ -240,7 +239,7 @@ public final class ModelMapGenMojo extends AbstractMojo {
             conf.put("process.class", fieldClass.getName());
             conf.put("process.date", ofLocalizedDateTime(SHORT).format(now()));
             conf.put("target.class.name", targetClassName);
-            conf.put("imports", imports(fieldInfoMap, null));
+            conf.put("imports", FieldInfoGen.imports(fieldInfoMap));
             conf.put("constants", constants(fieldInfoMap, enumFieldInfo));
             conf.put("source.generator.name", getClass().getName());
             final String content = MacroProcessor.replaceProperties(classTemplate, conf);
@@ -274,7 +273,7 @@ public final class ModelMapGenMojo extends AbstractMojo {
             conf.put("process.date", ofLocalizedDateTime(SHORT).format(now()));
             conf.put("target.class.name", targetClassName);
             conf.put("process.field.info.class", targetFieldInfoPackage + "." + fieldInfoClassName);
-            conf.put("imports", imports(fieldInfoMap, typeProvider));
+            conf.put("imports", DslMethodsGen.imports(fieldInfoMap, typeProvider));
             conf.put("methods", methods(fieldInfoMap, typeProvider, enumFieldInfo));
             conf.put("source.generator.name", getClass().getName());
             final String content = MacroProcessor.replaceProperties(Templates.dslFieldModel, conf);
