+++ generator/src/main/resources/io/doov/gen/DslFieldModel.template
@@ -1,15 +1,8 @@
 package ${package.name};
 
 import static ${process.field.info.class}.*;
-import static java.util.Collections.unmodifiableList;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import javax.annotation.Generated;
-
-import io.doov.core.dsl.field.*;
-import io.doov.core.*;
 
 ${imports}
