+++ generator/src/main/resources/io/doov/gen/FieldInfoEnum.template
@@ -3,15 +3,13 @@ package ${package.name};
 import static io.doov.core.dsl.field.FieldInfoBuilder.fieldInfo;
 import static java.util.Arrays.asList;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
 import javax.annotation.Generated;
 
-import io.doov.core.dsl.field.*;
-import io.doov.core.*;
+import io.doov.core.FieldInfo;
+import io.doov.core.dsl.field.DelegatingFieldInfo;
 
 ${imports}
